[backend] improve code, reorganize some methods, remove duplicates. (#5480)
