+++ openaev-api/src/main/java/io/openaev/config/ThreadPoolTaskLoggerConfig.java
@@ -24,6 +24,8 @@ public class ThreadPoolTaskLoggerConfig {
     private ThreadPoolTaskExecutor createBaseExecutor() {
         ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
 
+        //TODO: find a better way to configure this variables dynamically - maybe through properties file.
+
         executor.setCorePoolSize(10);
         executor.setMaxPoolSize(50);
         executor.setQueueCapacity(1000);
@@ -42,13 +44,16 @@ public class ThreadPoolTaskLoggerConfig {
             HttpServletRequest request = requestAttributes instanceof ServletRequestAttributes attrs ? attrs.getRequest() : null;
             Map<String, String> headers;
             String remoteAddress;
+            String method;
 
             if (request != null) {
                 headers = HttpReqRespUtils.extractHeaders(request);
                 remoteAddress = request.getRemoteAddr();
+                method = request.getMethod();
             } else {
                 headers = null;
                 remoteAddress = null;
+                method = null;
             }
 
             // CAPTURE LOGs CONTEXT
@@ -72,6 +77,7 @@ public class ThreadPoolTaskLoggerConfig {
                     // STORE HEADERS AND REMOTE ADDRESS
                     ThreadRequestContextHolder.setHeaders(headers);
                     ThreadRequestContextHolder.setRemoteAddress(remoteAddress);
+                    ThreadRequestContextHolder.setMethod(method);
 
                     // RESTORE MDC
                     if (mdcContext != null) {
@@ -140,6 +146,13 @@ public class ThreadPoolTaskLoggerConfig {
         }
         public static String getRemoteAddress() {
             return (String) get("REMOTE_ADDR");
+        }
+
+        public static void setMethod(String method) {
+            set("METHOD", method);
+        }
+        public static String getMethod() {
+            return (String) get("METHOD");
         }
 
         public static void clear() {
